print all correction

+++ 0x10-variadic_functions/3-print_all.c
@@ -30,7 +30,7 @@ void print_all(const char * const format, ...)
 				printed = 1;
 				break;
 			case 'f':
-				printf("%f", (float)va_arg(args, double));
+				printf("%f", va_arg(args, double));
 				printed = 1;
 				break;
 			case 's':
